retrieval/requirements.txt
minio
kafka-python
